Task Management Completed

+++ controllers/task-controller.js
@@ -15,6 +15,8 @@ const getTasks = (req, res) => {
   }
   if (type == 1) {
     return Task.find(query)
+      .populate("subtasks")
+      .populate("attachments")
       .sort({ created_at: -1 })
       .cursor()
       .pipe(JSONStream.stringify())
@@ -22,12 +24,16 @@ const getTasks = (req, res) => {
   }
   if (type == 2) {
     return Task.findDeleted(query)
+      .populate("subtasks")
+      .populate("attachments")
       .sort({ created_at: -1 })
       .cursor()
       .pipe(JSONStream.stringify())
       .pipe(res.type("json"))
   }
   return Task.findWithDeleted(query)
+    .populate("subtasks")
+    .populate("attachments")
     .sort({ created_at: -1 })
     .cursor()
     .pipe(JSONStream.stringify())
@@ -56,6 +62,7 @@ const createTask = async (req, res) => {
 
 const editTask = async (req, res) => {
   const { _id } = req.user
+  const { id } = req.params
   const {
     label,
     room,
@@ -69,32 +76,36 @@ const editTask = async (req, res) => {
     note
   } = req.body
 
+  if (!label) {
+    return res.status(400).json({ message: "Label is required" })
+  }
+
   if (progress && (progress < 0 || progress > 100)) {
     return res.status(400).json({ message: "Progress is between (0-100)" })
   }
 
-  // try {
-  await Task.updateOne(
-    { _id: req.params.id },
-    {
-      owner: _id,
-      label: label,
-      room,
-      depend_on: depend_on ? JSON.parse(depend_on) : undefined,
-      start_at,
-      end_at,
-      status,
-      priority,
-      location: JSON.parse(location),
-      progress,
-      note
-    }
-  )
-  const updatedTask = await Task.findById(req.params.id)
-  res.json(updatedTask)
-  // } catch (error) {
-  //   return res.status(500).json({ message: error })
-  // }
+  try {
+    const task = await Task.findById(id)
+      .populate("subtasks")
+      .populate("attachments")
+    if (!task) return res.status(400).json({ message: "Task not found" })
+    if (task.owner != _id && !task.assigned_to.includes(_id))
+      return res.status(400).json({ message: "Permission denined" })
+    task.label = label
+    task.room = room
+    task.depend_on = depend_on ? JSON.parse(depend_on) : undefined
+    task.start_at = start_at
+    task.end_at = end_at
+    task.status = status
+    task.priority = priority
+    task.location = location ? JSON.parse(location) : undefined
+    task.progress = progress
+    task.note = note
+    await task.save()
+    res.json(task)
+  } catch (error) {
+    return res.status(500).json({ message: error })
+  }
 }
 
 const assignTaskTo = async (req, res) => {
@@ -117,6 +128,8 @@ const assignTaskTo = async (req, res) => {
       { $addToSet: { assigned_to: JSON.parse(member_ids) } }
     )
     const updatedTask = await Task.findById(id)
+      .populate("subtasks")
+      .populate("attachments")
     res.json(updatedTask)
   } catch (error) {
     return res.status(500).json({ message: error })
@@ -143,6 +156,8 @@ const removeAssignTaskTo = async (req, res) => {
       { $pullAll: { assigned_to: JSON.parse(member_ids) } }
     )
     const updatedTask = await Task.findById(id)
+      .populate("subtasks")
+      .populate("attachments")
     res.json(updatedTask)
   } catch (error) {
     return res.status(500).json({ message: error })
@@ -150,8 +165,16 @@ const removeAssignTaskTo = async (req, res) => {
 }
 
 const deleteTask = async (req, res) => {
+  const { _id } = req.user
   try {
-    await Task.delete({ _id: req.params.id, owner: req.user._id })
+    const task = await Task.findById(req.params.id)
+    if (!task) {
+      return res.status(400).json({ message: "Task not found" })
+    }
+    if (task.owner != _id) {
+      return res.status(400).json({ message: "Permission denined" })
+    }
+    await Task.delete({ _id: req.params.id })
     return res.json({ message: "Delete successfully" })
   } catch (error) {
     return res.status(500).json({ message: error })
@@ -168,6 +191,10 @@ const addAttachment = async (req, res) => {
   }
 
   try {
+    const task = await Task.findById(id)
+    if (!task) return res.status(400).json({ message: "Task not found" })
+    if (task.owner != _id && !task.assigned_to.includes(_id))
+      return res.status(400).json({ message: "Permission denined" })
     if (Array.isArray(attachments)) {
       var resFiles = []
       for (let i = 0; i < attachments.length; i++) {
@@ -175,25 +202,17 @@ const addAttachment = async (req, res) => {
         const sss = await storeAttachment(element.path, id, _id)
         resFiles.push(sss)
       }
-      const newAttachments = await Attachment.insertMany(resFiles)
-      await Task.updateOne(
-        { _id: id, owner: _id },
-        {
-          $addToSet: { attachments: newAttachments }
-        }
-      )
+      await Attachment.insertMany(resFiles)
       const updatedData = await Task.findById(id)
+        .populate("subtasks")
+        .populate("attachments")
       return res.json(updatedData)
     } else {
       const resFile = await storeAttachment(attachments.path, id, _id)
-      const newAttachment = await Attachment.create(resFile)
-      await Task.updateOne(
-        { _id: id, owner: _id },
-        {
-          $addToSet: { attachments: [newAttachment._id] }
-        }
-      )
+      await Attachment.create(resFile)
       const updatedData = await Task.findById(id)
+        .populate("subtasks")
+        .populate("attachments")
       return res.json(updatedData)
     }
   } catch (error) {
@@ -235,16 +254,21 @@ const deleteAttachment = async (req, res) => {
   const { _id } = req.user
   const { id } = req.params
   const { attachment_ids } = req.body
+
   try {
+    const task = await Task.findById(id)
+      .populate("subtasks")
+      .populate("attachments")
+    if (!task) return res.status(400).json({ message: "Task not found" })
+    if (task.owner != _id && !task.assigned_to.includes(_id))
+      return res.status(400).json({ message: "Permission denined" })
     await Attachment.deleteMany({
       _id: { $in: JSON.parse(attachment_ids) },
       task: id
     })
-    await Task.updateOne(
-      { _id: id, owner: _id },
-      { $pullAll: { attachments: JSON.parse(attachment_ids) } }
-    )
     const updatedData = await Task.findById(id)
+      .populate("subtasks")
+      .populate("attachments")
     return res.json(updatedData)
   } catch (error) {
     return res.status(500).json({ error })
